Use pytest-asyncio instead of aiounittest
