+++ pytest.ini
@@ -1,0 +1,6 @@
+# pytest.ini
+[pytest]
+addopts = --strict-markers --tb=short -s
+testpaths = 
+  tests
+python_files = *.py
