+++ requirements-dev.txt
@@ -1,8 +1,8 @@
-aiounittest
 docutils
 mock
 Pygments
 pytest
+pytest-asyncio
 pytest-cov
 pytest-flake8
 Sphinx
